# experiment: Avoid copy when building from a single strict bytestring