@nikita-volkov 2022-06-16
Wait a minute.

I think a rewrite rule replacing `builderBytes . bytes` with `id` may solve all issues.
Actually, I think that it may be a nice example of what rewrite rules exist for.