@nikita-volkov 2022-06-17
I've added the rule in b9e69f54f3b0ced7216163685717cdd3c4595ccc and released it as 0.4.5.6. Unfortunately it won't affect the "copy" benchmark because the suite avoids inlining. Please check whether it changes anything at the postgrest end.

If it doesn't I think the best solution would be to implement something similar at the level of "postgresql-binary" [Encoding](https://hackage.haskell.org/package/postgresql-binary-0.12.4.3/docs/PostgreSQL-Binary-Encoding.html#t:Encoding). Namely to make it something like the following:

```
data Encoding
  = BuilderEncoding Builder
  | ByteStringEncoding ByteString
```

But first let's check whether rewrite rules have done any help.