@robx 2022-08-10
This doesn't seem to be the way forward, closing.